authority_server/server/requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
PyJWT==2.8.0
cryptography==41.0.4
requests==2.31.0
pycryptodome==3.18.0
